Pass exact serialised key algorithm string into isOK function
When faced with a simple algorithm string like e.g. "RSA" (which is all what we
get from querying PublicKey.getAlgorithm()), Android doesn't use the exact same
default cipher mode that gets used on Desktop (where the license was probably
originally created).

To fix that, we need to extract the original full algorithm specification string
stored in the serialised license key and pass that to the isOK function.